backend/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.23.2
sqlalchemy==2.0.22
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
pydantic-settings==2.0.0
pydantic==2.12.0
email-validator==2.1.1
numpy==1.26.2
faiss-cpu==1.7.4
httpx==0.26.0
sentence-transformers==2.2.2
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.2.2
huggingface-hub==0.16.4
prometheus_client==0.16.0
python-multipart==0.0.6
slowapi==0.1.5
alembic==1.11.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
PyPDF2==3.1.1
